+++ projects/igniteui-angular/src/lib/test-utils/template-strings.spec.ts
@@ -19,18 +19,19 @@ export class GridTemplateStrings {
     </igx-grid>`;
 
     public static declareGrid(attributes = ``, events = ``, columnDefinitions: ColumnDefinitions = ``,
-            toolbarDefinition = '', paginatorDefinition = '') {
+        toolbarDefinition = '', paginatorDefinition = '', templates: TemplateDefinitions = '') {
         return `<igx-grid [data]="data"
-        ${ attributes}
-        ${ events}
+        ${attributes}
+        ${events}
         >
-        ${ toolbarDefinition }
-        ${ columnDefinitions}
-        ${ paginatorDefinition }
+        ${toolbarDefinition}
+        ${columnDefinitions}
+        ${paginatorDefinition}
+        ${templates}
     </igx-grid>`;
     }
 
-    public static declareBasicGridWithColumns(columnDefinitions: ColumnDefinitions) {
+    public static declareBasicGridWithColumns(columnDefinitions: ColumnDefinitions, templates?: TemplateDefinitions) {
         return GridTemplateStrings.declareGrid(``, ``, columnDefinitions);
     }
 
@@ -486,6 +487,20 @@ export class ColumnDefinitions {
                 <igx-column field="PostalCode"></igx-column>
             </igx-column-group>
         </igx-column-group>
+    `;
+}
+
+export class TemplateDefinitions {
+    public static sortIconTemplates = `
+        <ng-template igxSortHeaderIcon>
+            <igx-icon>unfold_more</igx-icon>
+        </ng-template>
+        <ng-template igxSortAscendingHeaderIcon>
+            <igx-icon>expand_less</igx-icon>
+        </ng-template>
+        <ng-template igxSortDescendingHeaderIcon>
+            <igx-icon>expand_more</igx-icon>
+        </ng-template>
     `;
 }